+++ convertextract/__init__.py
@@ -1,5 +1,3 @@
 from .parsers import process, processText
 
-VERSION = "1.6.1"
-
-NAME = "convertextract"
+VERSION = "2.0"