+++ convertextract/cli.py
@@ -6,11 +6,15 @@ import argparse
 import encodings
 import os
 import pkgutil
+import sys
+import six
+import re
+import glob
 
 import argcomplete
 
 from . import VERSION
-from .parsers import DEFAULT_ENCODING
+from .parsers import DEFAULT_ENCODING, _get_available_extensions
 
 
 class AddToNamespaceAction(argparse.Action):
@@ -23,6 +27,17 @@ class AddToNamespaceAction(argparse.Action):
                 'Duplicate specification of the key "%(key)s" with --option.'
             ) % locals())
         setattr(namespace, key, val)
+
+
+# Fix FileType to honor 'b' flag, see: https://bugs.python.org/issue14156
+class FileType(argparse.FileType):
+    def __call__(self, string):
+        if string == '-' and six.PY3:
+            if 'r' in self._mode:
+                string = sys.stdin.fileno()
+            elif 'w' in self._mode:
+                string = sys.stdout.fileno()
+        return super(FileType, self).__call__(string)
 
 
 # This function is necessary to enable autodocumentation of the script
@@ -48,13 +63,19 @@ def get_parser():
         help='Specify the encoding of the output.',
     )
     parser.add_argument(
+        '--extension', type=str, default=None,
+        choices=_get_available_extensions(),
+        help='Specify the extension of the file.',
+    )
+    parser.add_argument(
         '-m', '--method', default='',
         help='Specify a method of extraction for formats that support it',
     )
     parser.add_argument(
-        '-o', '--output', type=argparse.FileType('w'), default='-',
+        '-o', '--output', type=FileType('wb'), default='-',
         help='Output raw text in this file',
     )
+    parser.add_argument('--no-write', dest='no_write', action='store_true', help="Disable default writing of converted file.")
     parser.add_argument(
         '-l', '--language', type=str,
         help='Specify language for conversion. Can be either user defined'
@@ -62,18 +83,13 @@ def get_parser():
             ' list. For a full list please visit https://github.com/roedoejet/convertextract/',
     )
     parser.add_argument(
-        '-p', '--path', type=str,
-        help='Specify path for conversion'
-              'for multiple paths, must be comma-delimited string',
+        '-O', '--option', type=str, action=AddToNamespaceAction,
+        help=(
+            'Add arbitrary options to various parsers of the form '
+            'KEYWORD=VALUE. A full list of available KEYWORD options is '
+            'available at http://bit.ly/textract-options'
+        ),
     )
-#    parser.add_argument(
-#        '-O', '--option', type=str, action=AddToNamespaceAction,
-#        help=(
-#            'Add arbitrary options to various parsers of the form '
-#            'KEYWORD=VALUE. A full list of available KEYWORD options is '
-#            'available at http://bit.ly/textract-options'
-#        ),
-#    )
     parser.add_argument(
         '-v', '--version', action='version', version='%(prog)s '+VERSION,
     )