+++ convertextract/exceptions.py
@@ -10,19 +10,6 @@ class CommandLineError(Exception):
     def render(self, msg):
         return msg % vars(self)
 
-
-class ExtensionNotSupported(CommandLineError):
-    """This error is raised with unsupported extensions"""
-    def __init__(self, ext):
-        self.ext = ext
-
-    def __str__(self):
-        return self.render((
-            'The filename extension %(ext)s is not yet supported by\n'
-            'textract. Please suggest this filename extension here:\n\n'
-            '    https://github.com/roedoejet/convertextract/issues\n'
-        ))
-    
 class CorrespondenceMissing(CommandLineError):
     def __init__(self, language):
         self.language = language
@@ -32,6 +19,27 @@ class CorrespondenceMissing(CommandLineError):
             'There is no correspondence with the name "%(language)s", please\n'
             'make sure you spelled the name correctly or go to\n'
             'https://github.com/roedoejet/convertextract/ for a list of correspondences'
+        ))
+
+
+class ExtensionNotSupported(CommandLineError):
+    """This error is raised with unsupported extensions"""
+    def __init__(self, ext):
+        self.ext = ext
+
+        from .parsers import _get_available_extensions
+        available_extensions = []
+        for e in _get_available_extensions():
+            if e.startswith('.'):
+                available_extensions.append(e)
+        self.available_extensions_str = ', '.join(available_extensions)
+
+    def __str__(self):
+        return self.render((
+            'The filename extension %(ext)s is not yet supported by\n'
+            'convertextract. Please suggest this filename extension here:\n\n'
+            '    https://github.com/roedoeojet/convertextract/issues\n\n'
+            'Available extensions include: %(available_extensions_str)s\n'
         ))
 
 
@@ -46,7 +54,7 @@ class MissingFileError(CommandLineError):
     def __str__(self):
         return self.render((
             'The file "%(filename)s" can not be found.\n'
-            'Is this the right path/to/file/you/want/to/convert%(ext)s?'
+            'Is this the right path/to/file/you/want/to/extract%(ext)s?'
         ))
 
 
@@ -83,7 +91,6 @@ class ShellError(CommandLineError):
             "`%(executable)s` is not installed on your system. Please make\n"
             "sure the appropriate dependencies are installed before using\n"
             "convertextract:\n\n"
-            "    http://textract.readthedocs.org/en/latest/installation.html\n"
         ) % vars(self)
 
     def failed_message(self):