+++ convertextract/parsers/__init__.py
@@ -4,15 +4,20 @@ Route the request to the appropriate parser based on file type.
 
 import os
 import importlib
+import glob
+import re
 
 from .. import exceptions
 from ..cors import processCors
 
 # Dictionary structure for synonymous file extension types
 EXTENSION_SYNONYMS = {
-    ".jpeg": ".jpg",
+    # ".jpeg": ".jpg",
+    # ".tff": ".tiff",
+    # ".tif": ".tiff",
     ".htm": ".html",
     "": ".txt",
+    ".log": ".txt",
 }
 
 # default encoding that is returned by the process method. specify it
@@ -20,20 +25,16 @@ EXTENSION_SYNONYMS = {
 # the command line interface
 DEFAULT_ENCODING = 'utf_8'
 
+# filename format
+_FILENAME_SUFFIX = '_parser'
 
-def process(filename, encoding=DEFAULT_ENCODING, **kwargs):
+
+def process(filename, encoding=DEFAULT_ENCODING, extension=None, **kwargs):
     """This is the core function used for extracting text. It routes the
     ``filename`` to the appropriate parser and returns the extracted
     text as a byte-string encoded with ``encoding``.
     """
-    
-    # make sure optional kwargs are None is not supplied
-    if not "language" in kwargs:
-        kwargs["language"] = None
-    
-    if not "path" in kwargs:
-        kwargs["path"] = None
-    
+
     # make sure the filename exists
     if not os.path.exists(filename):
         raise exceptions.MissingFileError(filename)
@@ -41,8 +42,18 @@ def process(filename, encoding=DEFAULT_ENCODING, **kwargs):
     # get the filename extension, which is something like .docx for
     # example, and import the module dynamically using importlib. This
     # is a relative import so the name of the package is necessary
-    _, ext = os.path.splitext(filename)
-    ext = ext.lower()
+    # normally, file extension will be extracted from the file name
+    # if the file name has no extension, then the user can pass the
+    # extension as an argument
+    if extension:
+        ext = extension
+        # check if the extension has the leading .
+        if not ext.startswith('.'):
+            ext = '.' + ext
+        ext = ext.lower()
+    else:
+        _, ext = os.path.splitext(filename)
+        ext = ext.lower()
 
     # check the EXTENSION_SYNONYMS dictionary
     ext = EXTENSION_SYNONYMS.get(ext, ext)
@@ -50,18 +61,23 @@ def process(filename, encoding=DEFAULT_ENCODING, **kwargs):
     # to avoid conflicts with packages that are installed globally
     # (e.g. python's json module), all extension parser modules have
     # the _parser extension
-    rel_module = ext + '_parser'
-    module_name = rel_module[1:]
+    rel_module = ext + _FILENAME_SUFFIX
 
-    # if this module name doesn't exist in this directory it isn't
-    # currently supported
-    this_dir = os.path.dirname(os.path.abspath(__file__))
-    if not os.path.exists(os.path.join(this_dir, module_name + '.py')):
+    # If we can't import the module, the file extension isn't currently
+    # supported
+    try:
+        filetype_module = importlib.import_module(
+            rel_module, 'convertextract.parsers'
+        )
+    except ImportError:
         raise exceptions.ExtensionNotSupported(ext)
 
     # do the extraction
-    filetype_module = importlib.import_module(rel_module, 'convertextract.parsers')
-    parser = filetype_module.Parser()
+
+    parser = filetype_module.Parser(**kwargs)
+    # print("encoding is " + encoding)
+    # print(type(parser.process(filename, encoding, **kwargs)))
+    # print(parser.process(filename, encoding, **kwargs))
     return parser.process(filename, encoding, **kwargs)
 
 def processText(text, **kwargs):
@@ -76,4 +92,26 @@ def processText(text, **kwargs):
     cors = processCors(kwargs["language"])
 
     return cors.apply_rules(text)
+
+
+def _get_available_extensions():
+    """Get a list of available file extensions to make it easy for
+    tab-completion and exception handling.
+    """
+    extensions = []
     
+    # from filenames
+    parsers_dir = os.path.join(os.path.dirname(__file__))
+    glob_filename = os.path.join(parsers_dir, "*" + _FILENAME_SUFFIX + ".py")
+    for filename in glob.glob(glob_filename):
+        ext = filename.split("_")[0]
+        extensions.append(ext)
+        extensions.append('.' + ext)
+
+    # from relevant synonyms (don't use the '' synonym)
+    for ext in EXTENSION_SYNONYMS.keys():
+        if ext:
+            extensions.append(ext)
+            extensions.append(ext.replace('.', '', 1))
+    extensions.sort()
+    return extensions