+++ convertextract/parsers/utils.py
@@ -5,7 +5,9 @@ reused in many of the other parser modules.
 import subprocess
 import tempfile
 import os
+import errno
 
+import six
 import chardet
 
 from .. import exceptions
@@ -16,7 +18,9 @@ class BaseParser(object):
     that is used across all document Parsers. In particular, it has
     the responsibility of handling all unicode and byte-encoding.
     """
-
+    def __init__(self, **kwargs): 
+        pass
+        
     def extract(self, filename, **kwargs):
         """This method must be overwritten by child classes to extract raw
         text from a filename. This method can return either a
@@ -32,7 +36,7 @@ class BaseParser(object):
 
     def process(self, filename, encoding, **kwargs):
         """Process ``filename`` and encode byte-string with ``encoding``. This
-        method is called by :func:`textract.parsers.process` and wraps
+        method is called by :func:`convertextract.parsers.process` and wraps
         the :meth:`.BaseParser.extract` method in `a delicious unicode
         sandwich <http://nedbatchelder.com/text/unipain.html>`_.
 
@@ -51,7 +55,7 @@ class BaseParser(object):
         """
         # only decode byte strings into unicode if it hasn't already
         # been done by a subclass
-        if isinstance(text, unicode):
+        if isinstance(text, six.text_type):
             return text
 
         # empty text? nothing to decode
@@ -59,7 +63,6 @@ class BaseParser(object):
             return u''
 
         # use chardet to automatically detect the encoding text
-        max_confidence, max_encoding = 0.0, None
         result = chardet.detect(text)
         return text.decode(result['encoding'])
 
@@ -70,17 +73,25 @@ class ShellParser(BaseParser):
     `Fabric <http://www.fabfile.org/>`_-like behavior.
     """
 
-    def run(self, command):
+    def run(self, args):
         """Run ``command`` and return the subsequent ``stdout`` and ``stderr``
         as a tuple. If the command is not successful, this raises a
-        :exc:`textract.exceptions.ShellError`.
+        :exc:`convertextract.exceptions.ShellError`.
         """
 
         # run a subprocess and put the stdout and stderr on the pipe object
-        pipe = subprocess.Popen(
-            command, shell=True,
-            stdout=subprocess.PIPE, stderr=subprocess.PIPE,
-        )
+        try:
+            pipe = subprocess.Popen(
+                args,
+                stdout=subprocess.PIPE, stderr=subprocess.PIPE,
+            )
+        except OSError as e:
+            if e.errno == errno.ENOENT:
+                # File not found.
+                # This is equivalent to getting exitcode 127 from sh
+                raise exceptions.ShellError(
+                    ' '.join(args), 127, '', '',
+                )
 
         # pipe.wait() ends up hanging on large files. using
         # pipe.communicate appears to avoid this issue
@@ -89,7 +100,7 @@ class ShellParser(BaseParser):
         # if pipe is busted, raise an error (unlike Fabric)
         if pipe.returncode != 0:
             raise exceptions.ShellError(
-                command, pipe.returncode, stdout, stderr,
+                ' '.join(args), pipe.returncode, stdout, stderr,
             )
 
         return stdout, stderr